requirements.txt
agent-framework-foundry
azure-identity
python-dotenv
